requirements.txt
streamlit==1.32.2
openai==1.55.3
httpx==0.27.2
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
python-dotenv==1.0.1
